Refactor rule_validator
